requirements.txt
asgiref==3.8.1
charset-normalizer==3.3.2
dj-database-url==2.2.0
Django==4.2.15
django-gravatar==0.1.0
django-gravatar2==1.4.4
djangorestframework==3.15.2


